// src/app/models/category.enum.ts
export enum CategoryEnum {
  TELEVISION = 'TELEVISION',
  AUDIO = 'AUDIO',
  COMPUTERS = 'COMPUTERS',
  APPLIANCES = 'APPLIANCES',
  MOBILE_PHONES = 'MOBILE_PHONES',
  GAMING = 'GAMING',
  WEARABLES = 'WEARABLES',
  CAMERAS = 'CAMERAS',
  HOME_AUTOMATION = 'HOME_AUTOMATION',
  HOME_APPLIANCES = 'HOME_APPLIANCES',
  PERSONAL_CARE = 'PERSONAL_CARE',
  OUTDOOR = 'OUTDOOR',
  SECURITY = 'SECURITY',
  ACCESSORIES = 'ACCESSORIES',
  MUSICAL_INSTRUMENTS = 'MUSICAL_INSTRUMENTS',
  ELECTRONICS = 'ELECTRONICS',
  TABLETS = 'TABLETS',
}